Refs #21095: Fix casting for windows compilation

+++ examples/cpp/discovery_server/CLIParser.hpp
@@ -500,7 +500,7 @@ public:
                             }
                             else
                             {
-                                config.pub_config.interval = static_cast<uint32_t>(input);
+                                config.pub_config.interval = static_cast<uint16_t>(input);
                             }
                         }
                         catch (const std::invalid_argument& e)
